requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0
mysqlclient>=2.2,<3.0
mysql-connector-python>=8.0,<9.0
python-dotenv>=1.0,<2.0
